Added error handling to imageStream, modified saveFile to handle naming
Errors thrown by the stream are caught to prevent the error coming
from line 94 in streams.js.
saveFile was modified so that it takes an object instead of a filename.
The object contains the user.name and id_str. The saveFile function
itself creates the filename string from the user.name,the id_str and the
image name from the url.